refactor(interactive-shell): modular planner, executor, registry
- Add intent_parser, action_planner, action_executor, rendering, interaction_models

- Add command_registry package with composable COMMANDS and repl_data loaders

- Slim agent_actions/commands; update tests for patch targets (repl_data, executor)
