+++ app/cli/interactive_shell/commands.py
@@ -1,958 +1,19 @@
-"""Slash command handlers for the REPL."""
+"""Slash command handlers for the REPL (compatibility shim for the modular registry)."""
 
 from __future__ import annotations
 
-import os
-from collections.abc import Callable
-from dataclasses import dataclass
-from datetime import UTC, datetime
-from typing import Any
-
-from rich import box
-from rich.console import Console
-from rich.markup import escape
-from rich.table import Table
-
-from app.cli.interactive_shell.banner import render_banner, resolve_provider_models
-from app.cli.interactive_shell.history import load_command_history_entries
-from app.cli.interactive_shell.session import ReplSession
-from app.cli.interactive_shell.tasks import TaskKind, TaskRecord, TaskStatus
-from app.cli.interactive_shell.theme import TERMINAL_ACCENT_BOLD, TERMINAL_ERROR
-from app.cli.support.errors import OpenSREError
-from app.utils.sentry_sdk import capture_exception
-
-
-def _repl_table(**kwargs: Any) -> Table:
-    """Minimal outer borders — closer to Claude Code than full ASCII grids."""
-    opts: dict[str, Any] = {
-        "box": box.MINIMAL_HEAVY_HEAD,
-        "show_edge": False,
-        "pad_edge": False,
-    }
-    opts.update(kwargs)
-    return Table(**opts)
-
-
-@dataclass(frozen=True)
-class SlashCommand:
-    name: str
-    help_text: str
-    handler: Callable[[ReplSession, Console, list[str]], bool]
-
-
-def _cmd_help(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    table = _repl_table(title="Slash commands", title_style=TERMINAL_ACCENT_BOLD, show_header=False)
-    table.add_column("name", style="bold")
-    table.add_column("description", style="dim")
-    for cmd in SLASH_COMMANDS.values():
-        table.add_row(cmd.name, cmd.help_text)
-    console.print(table)
-    return True
-
-
-def _cmd_exit(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    console.print("[dim]goodbye.[/dim]")
-    return False
-
-
-def _cmd_clear(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    console.clear()
-    render_banner(console)
-    return True
-
-
-def _cmd_reset(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    session.clear()
-    console.print("[dim]session state cleared.[/dim]")
-    return True
-
-
-def _cmd_trust(session: ReplSession, console: Console, args: list[str]) -> bool:
-    if args and args[0].lower() in ("off", "false", "disable"):
-        session.trust_mode = False
-        console.print("[dim]trust mode off[/dim]")
-    else:
-        session.trust_mode = True
-        console.print("[yellow]trust mode on[/yellow] — future approval prompts will be skipped")
-    return True
-
-
-def _cmd_status(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    table = _repl_table(title="Session status", title_style=TERMINAL_ACCENT_BOLD, show_header=False)
-    table.add_column("key", style="bold")
-    table.add_column("value")
-    table.add_row("interactions", str(len(session.history)))
-    table.add_row("last investigation", "yes" if session.last_state else "none")
-    table.add_row("trust mode", "on" if session.trust_mode else "off")
-    table.add_row("provider", os.getenv("LLM_PROVIDER", "anthropic"))
-    acc = session.accumulated_context
-    if acc:
-        table.add_row("accumulated context", ", ".join(sorted(acc.keys())))
-    console.print(table)
-    return True
-
-
-# MCP-type services are rendered separately under `/list mcp` so the default
-# `/list integrations` view stays focused on alert-source / data integrations.
-_MCP_SERVICES = frozenset({"github", "openclaw"})
-
-
-def _load_verified_integrations() -> list[dict[str, str]]:
-    """Import lazily so an unconfigured store doesn't slow down every REPL turn."""
-    from app.integrations.verify import verify_integrations
-
-    return verify_integrations()
-
-
-def _load_llm_settings() -> Any | None:
-    """Best-effort LLM settings load; returns None if env is misconfigured."""
-    try:
-        from app.config import LLMSettings
-
-        return LLMSettings.from_env()
-    except Exception:  # noqa: BLE001 — env/config errors are expected here
-        return None
-
-
-def _status_style(status: str) -> str:
-    return {
-        "ok": "green",
-        "configured": "green",
-        "missing": "yellow",
-        "failed": "red",
-    }.get(status, "dim")
-
-
-def _render_integrations_table(console: Console, results: list[dict[str, str]]) -> None:
-    rows = [r for r in results if r.get("service") not in _MCP_SERVICES]
-    if not rows:
-        console.print("[dim]no integrations configured.  try `opensre onboard` to add one.[/dim]")
-        return
-    table = _repl_table(title="Integrations", title_style=TERMINAL_ACCENT_BOLD)
-    table.add_column("service", style="bold")
-    table.add_column("source", style="dim")
-    table.add_column("status")
-    table.add_column("detail", style="dim", overflow="fold")
-    for row in rows:
-        status = row.get("status", "unknown")
-        table.add_row(
-            escape(row.get("service", "?")),
-            escape(row.get("source", "?")),
-            f"[{_status_style(status)}]{escape(status)}[/{_status_style(status)}]",
-            escape(row.get("detail", "")),
-        )
-    console.print(table)
-
-
-def _render_mcp_table(console: Console, results: list[dict[str, str]]) -> None:
-    rows = [r for r in results if r.get("service") in _MCP_SERVICES]
-    if not rows:
-        console.print("[dim]no MCP servers configured.[/dim]")
-        return
-    table = _repl_table(title="MCP servers", title_style=TERMINAL_ACCENT_BOLD)
-    table.add_column("server", style="bold")
-    table.add_column("source", style="dim")
-    table.add_column("status")
-    table.add_column("detail", style="dim", overflow="fold")
-    for row in rows:
-        status = row.get("status", "unknown")
-        table.add_row(
-            escape(row.get("service", "?")),
-            escape(row.get("source", "?")),
-            f"[{_status_style(status)}]{escape(status)}[/{_status_style(status)}]",
-            escape(row.get("detail", "")),
-        )
-    console.print(table)
-
-
-def _render_models_table(console: Console) -> None:
-    settings = _load_llm_settings()
-    if settings is None:
-        console.print(f"[{TERMINAL_ERROR}]LLM settings unavailable[/] — check provider env vars.")
-        return
-    provider = str(getattr(settings, "provider", "unknown"))
-    reasoning_model, toolcall_model = resolve_provider_models(settings, provider)
-    table = _repl_table(title="LLM connection", title_style=TERMINAL_ACCENT_BOLD, show_header=False)
-    table.add_column("key", style="bold")
-    table.add_column("value")
-    table.add_row("provider", provider)
-    table.add_row("reasoning model", reasoning_model)
-    table.add_row("toolcall model", toolcall_model)
-    console.print(table)
-
-
-def switch_llm_provider(
-    provider_name: str,
-    console: Console,
-    model: str | None = None,
-    *,
-    toolcall_model: str | None = None,
-) -> bool:
-    from app.cli.wizard.config import PROVIDER_BY_VALUE
-    from app.cli.wizard.env_sync import sync_env_values
-    from app.llm_credentials import has_llm_api_key
-
-    provider_key = provider_name.strip().lower()
-    provider = PROVIDER_BY_VALUE.get(provider_key)
-    if provider is None:
-        choices = ", ".join(sorted(PROVIDER_BY_VALUE))
-        console.print(
-            f"[{TERMINAL_ERROR}]unknown LLM provider:[/] {escape(provider_name)} "
-            f"[dim](choices: {choices})[/dim]"
-        )
-        return False
-
-    # Refuse to half-update .env when the target provider has no usable
-    # credential. Without this the user lands in a state where LLM_PROVIDER
-    # points at e.g. anthropic but ANTHROPIC_API_KEY is unset, so the very
-    # next call into LLMSettings.from_env() raises and /model show prints
-    # "LLM settings unavailable" — which is exactly what reviewers caught
-    # in #1192. Skip the check for providers whose credential isn't a
-    # secret (ollama uses OLLAMA_HOST which has a working default) and for
-    # CLI-backed providers (codex, claude-code) that authenticate through
-    # the vendor CLI and have no api_key_env at all.
-    if (
-        provider.credential_secret
-        and provider.api_key_env
-        and not has_llm_api_key(provider.api_key_env)
-    ):
-        console.print(
-            f"[{TERMINAL_ERROR}]missing credential for {provider.value}:[/] "
-            f"{provider.api_key_env} is not set in env or the keyring."
-        )
-        console.print(
-            f"[dim]set it with[/dim] [bold]export {provider.api_key_env}=<your-key>[/bold] "
-            "[dim]or run[/dim] [bold]opensre onboard[/bold] "
-            "[dim]to save it to the keyring, then rerun this command.[/dim]"
-        )
-        return False
-
-    selected_model = model.strip() if model else provider.default_model
-    values = {"LLM_PROVIDER": provider.value, provider.model_env: selected_model}
-    if provider.legacy_model_env:
-        values[provider.legacy_model_env] = selected_model
-
-    selected_toolcall: str | None = None
-    if toolcall_model is not None:
-        if not provider.toolcall_model_env:
-            console.print(
-                f"[yellow]provider {provider.value} does not expose a separate "
-                "toolcall model[/yellow] — toolcall override ignored."
-            )
-        else:
-            selected_toolcall = toolcall_model.strip()
-            if selected_toolcall:
-                values[provider.toolcall_model_env] = selected_toolcall
-
-    env_path = sync_env_values(values)
-    os.environ.update(values)
-
-    # Be explicit about which slot each model lands in. The previous output
-    # ("switched LLM provider: anthropic (X)") read ambiguously: a reviewer
-    # ran `/model set anthropic claude-haiku-...` and asked "how is the
-    # reasoning model changing then?" because `(X)` did not say *which*
-    # slot X went into. Always label both slots on a switch.
-    console.print(f"[green]switched LLM provider:[/green] {provider.value}")
-    console.print(
-        f"[green]reasoning model:[/green] {selected_model or 'provider default'} "
-        f"[dim]({provider.model_env})[/dim]"
-    )
-    if selected_toolcall:
-        console.print(
-            f"[green]toolcall model:[/green] {selected_toolcall} "
-            f"[dim]({provider.toolcall_model_env})[/dim]"
-        )
-    console.print(f"[dim]updated {env_path}[/dim]")
-    _render_models_table(console)
-    return True
-
-
-def switch_toolcall_model(
-    toolcall_model: str,
-    console: Console,
-    *,
-    provider_name: str | None = None,
-) -> bool:
-    """Set the toolcall model for the active (or named) provider."""
-    from app.cli.wizard.config import PROVIDER_BY_VALUE
-    from app.cli.wizard.env_sync import sync_env_values
-
-    raw_name = provider_name if provider_name else os.getenv("LLM_PROVIDER", "anthropic")
-    resolved_name = (raw_name or "anthropic").strip().lower()
-    provider = PROVIDER_BY_VALUE.get(resolved_name)
-    if provider is None:
-        choices = ", ".join(sorted(PROVIDER_BY_VALUE))
-        console.print(
-            f"[{TERMINAL_ERROR}]unknown LLM provider:[/] {escape(resolved_name)} "
-            f"[dim](choices: {choices})[/dim]"
-        )
-        return False
-    if not provider.toolcall_model_env:
-        console.print(
-            f"[yellow]provider {provider.value} does not expose a separate "
-            "toolcall model[/yellow] — nothing to set."
-        )
-        return False
-    new_model = toolcall_model.strip()
-    if not new_model:
-        console.print(f"[{TERMINAL_ERROR}]toolcall model cannot be empty[/]")
-        return False
-
-    values = {provider.toolcall_model_env: new_model}
-    env_path = sync_env_values(values)
-    os.environ.update(values)
-
-    console.print(
-        f"[green]toolcall model set to:[/green] {new_model} "
-        f"[dim]({provider.value} · {provider.toolcall_model_env})[/dim]"
-    )
-    console.print(f"[dim]updated {env_path}[/dim]")
-    _render_models_table(console)
-    return True
-
-
-def _parse_model_set_args(args: list[str]) -> tuple[str, str | None, str | None]:
-    """Parse `set <provider> [reasoning_model] [--toolcall-model <m>]`.
-
-    ``args`` is the slice after the ``set``/``use``/``switch`` keyword.
-    Raises :class:`ValueError` with a user-facing message when the input is
-    malformed, so the caller can surface a specific reason ("missing value
-    for --toolcall-model") instead of a generic usage line.
-    """
-    if not args:
-        raise ValueError("missing provider name")
-
-    provider = args[0]
-    reasoning_model: str | None = None
-    toolcall_model: str | None = None
-
-    i = 1
-    while i < len(args):
-        token = args[i]
-        if token == "--toolcall-model":
-            if i + 1 >= len(args):
-                raise ValueError("missing value for --toolcall-model")
-            toolcall_model = args[i + 1]
-            i += 2
-            continue
-        if token.startswith("--"):
-            raise ValueError(f"unknown flag: {token}")
-        if reasoning_model is not None:
-            raise ValueError(f"unexpected extra argument: {token}")
-        reasoning_model = token
-        i += 1
-
-    return provider, reasoning_model, toolcall_model
-
-
-def _cmd_integrations(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    sub = (args[0].lower() if args else "list").strip()
-
-    if sub in ("list", "ls"):
-        _render_integrations_table(console, _load_verified_integrations())
-        return True
-
-    if sub == "verify":
-        results = _load_verified_integrations()
-        _render_integrations_table(console, results)
-        failed = [r for r in results if r.get("status") in ("failed", "missing")]
-        if failed:
-            console.print(f"[yellow]{len(failed)} integration(s) need attention.[/yellow]")
-        else:
-            console.print("[green]all integrations ok.[/green]")
-        return True
-
-    if sub == "show":
-        if len(args) < 2:
-            console.print("[dim]usage:[/dim] /integrations show <service>")
-            return True
-        service = args[1].lower()
-        results = _load_verified_integrations()
-        match = next((r for r in results if r.get("service") == service), None)
-        if match is None:
-            console.print(f"[{TERMINAL_ERROR}]service not found:[/] {escape(service)}")
-            return True
-        table = _repl_table(
-            title=f"Integration: {service}",
-            title_style=TERMINAL_ACCENT_BOLD,
-            show_header=False,
-        )
-        table.add_column("key", style="bold")
-        table.add_column("value")
-        for k, v in match.items():
-            table.add_row(k, str(v))
-        console.print(table)
-        return True
-
-    console.print(
-        f"[{TERMINAL_ERROR}]unknown subcommand:[/] {escape(sub)}  "
-        "(try [bold]/integrations list[/bold], [bold]/integrations verify[/bold], "
-        "or [bold]/integrations show <service>[/bold])"
-    )
-    return True
-
-
-def _cmd_mcp(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    sub = (args[0].lower() if args else "list").strip()
-
-    if sub in ("list", "ls"):
-        _render_mcp_table(console, _load_verified_integrations())
-        return True
-
-    if sub == "connect":
-        console.print(
-            "[dim]to connect an MCP server, run:[/dim] [bold]opensre integrations setup[/bold]"
-        )
-        return True
-
-    if sub == "disconnect":
-        console.print(
-            "[dim]to remove an MCP server, run:[/dim] [bold]opensre integrations remove <service>[/bold]"
-        )
-        return True
-
-    console.print(
-        f"[{TERMINAL_ERROR}]unknown subcommand:[/] {escape(sub)}  "
-        "(try [bold]/mcp list[/bold], [bold]/mcp connect[/bold], or [bold]/mcp disconnect[/bold])"
-    )
-    return True
-
-
-def _cmd_model(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    sub = (args[0].lower() if args else "show").strip()
-
-    if sub == "show":
-        _render_models_table(console)
-        return True
-
-    if sub == "toolcall":
-        # /model toolcall set <model>   — set toolcall model for active provider
-        # /model toolcall show          — alias for /model show
-        if len(args) >= 2 and args[1].lower() == "show":
-            _render_models_table(console)
-            return True
-        if len(args) >= 2 and args[1].lower() in ("set", "use", "switch"):
-            if len(args) < 3:
-                console.print("[dim]usage:[/dim] /model toolcall set <model>")
-                return True
-            switch_toolcall_model(args[2], console)
-            return True
-        console.print(
-            "[dim]usage:[/dim] /model toolcall set <model> "
-            "[dim](sets the toolcall model for the active provider)[/dim]"
-        )
-        return True
-
-    if sub in ("set", "use", "switch"):
-        try:
-            provider_name, reasoning_model, toolcall_model = _parse_model_set_args(args[1:])
-        except ValueError as exc:
-            console.print()
-            console.print(f"[{TERMINAL_ERROR}]{escape(str(exc))}[/]")
-            console.print()
-            console.print(
-                "[dim]usage:[/dim] /model set <provider> [model] [--toolcall-model <model>]"
-            )
-            return True
-        switch_llm_provider(
-            provider_name,
-            console,
-            model=reasoning_model,
-            toolcall_model=toolcall_model,
-        )
-        return True
-
-    console.print(
-        f"[{TERMINAL_ERROR}]unknown subcommand:[/] {escape(sub)}  "
-        "(try [bold]/model show[/bold], "
-        "[bold]/model set <provider> [model] [--toolcall-model <m>][/bold], "
-        "or [bold]/model toolcall set <model>[/bold])"
-    )
-    return True
-
-
-def _cmd_health(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    from app.cli.support.health_view import render_health_report
-    from app.config import get_environment
-    from app.integrations.store import STORE_PATH
-    from app.integrations.verify import verify_integrations
-
-    results = verify_integrations()
-    environment = get_environment().value
-    render_health_report(
-        console=console,
-        environment=environment,
-        integration_store_path=STORE_PATH,
-        results=results,
-    )
-    return True
-
-
-def _cmd_doctor(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    from app.cli.commands.doctor import _CHECKS, _check
-
-    _STATUS_STYLES: dict[str, str] = {"ok": "green", "warn": "yellow", "error": "red"}
-    table = _repl_table(title="OpenSRE Doctor", title_style=TERMINAL_ACCENT_BOLD)
-    table.add_column("check", style="bold")
-    table.add_column("status")
-    table.add_column("detail", style="dim", overflow="fold")
-
-    issues = 0
-    for name, fn in _CHECKS:
-        result = _check(name, fn)
-        status = result["status"]
-        style = _STATUS_STYLES.get(status, "dim")
-        table.add_row(name, f"[{style}]{status}[/{style}]", result["detail"])
-        if status in ("warn", "error"):
-            issues += 1
-
-    console.print(table)
-    if issues:
-        console.print(f"[yellow]{issues} issue(s) found.[/yellow]")
-    else:
-        console.print("[green]all checks passed.[/green]")
-    return True
-
-
-def _cmd_version(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    import platform
-
-    from app.version import get_version
-
-    table = _repl_table(title="Version info", title_style=TERMINAL_ACCENT_BOLD, show_header=False)
-    table.add_column("key", style="bold")
-    table.add_column("value")
-    table.add_row("opensre", get_version())
-    table.add_row("python", platform.python_version())
-    table.add_row("os", f"{platform.system().lower()} ({platform.machine()})")
-    console.print(table)
-    return True
-
-
-def _cmd_template(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    import json
-
-    from app.cli.investigation.alert_templates import build_alert_template
-    from app.cli.support.constants import ALERT_TEMPLATE_CHOICES
-
-    if not args:
-        console.print(
-            f"[dim]usage:[/dim] /template <type>  (choices: {', '.join(ALERT_TEMPLATE_CHOICES)})"
-        )
-        return True
-
-    template_name = args[0].lower()
-    try:
-        payload = build_alert_template(template_name)
-    except ValueError:
-        console.print(
-            f"[{TERMINAL_ERROR}]unknown template:[/] {escape(template_name)}  "
-            f"(choices: {', '.join(ALERT_TEMPLATE_CHOICES)})"
-        )
-        return True
-
-    console.print_json(json.dumps(payload, indent=2))
-    return True
-
-
-def _cmd_investigate_file(session: ReplSession, console: Console, args: list[str]) -> bool:
-    from pathlib import Path
-
-    from app.cli.investigation import run_investigation_for_session
-
-    if not args:
-        console.print("[dim]usage:[/dim] /investigate <file>")
-        return True
-
-    path = Path(args[0])
-    if not path.exists():
-        console.print(f"[{TERMINAL_ERROR}]file not found:[/] {escape(str(path))}")
-        return True
-
-    try:
-        text = path.read_text(encoding="utf-8")
-    except Exception as exc:  # noqa: BLE001
-        console.print(f"[{TERMINAL_ERROR}]cannot read file:[/] {escape(str(exc))}")
-        return True
-
-    task = session.task_registry.create(TaskKind.INVESTIGATION)
-    task.mark_running()
-    try:
-        final_state = run_investigation_for_session(
-            alert_text=text,
-            context_overrides=session.accumulated_context or None,
-            cancel_requested=task.cancel_requested,
-        )
-    except KeyboardInterrupt:
-        task.mark_cancelled()
-        console.print("[yellow]investigation cancelled.[/yellow]")
-        session.record("alert", args[0], ok=False)
-        return True
-    except OpenSREError as exc:
-        task.mark_failed(str(exc))
-        console.print(f"[{TERMINAL_ERROR}]investigation failed:[/] {escape(str(exc))}")
-        if exc.suggestion:
-            console.print(f"[yellow]suggestion:[/yellow] {escape(exc.suggestion)}")
-        session.record("alert", args[0], ok=False)
-        return True
-    except Exception as exc:  # noqa: BLE001
-        task.mark_failed(str(exc))
-        capture_exception(exc)
-        console.print(f"[{TERMINAL_ERROR}]investigation failed:[/] {escape(str(exc))}")
-        session.record("alert", args[0], ok=False)
-        return True
-
-    root = final_state.get("root_cause")
-    task.mark_completed(result=str(root) if root is not None else "")
-    session.last_state = final_state
-    # Match `_run_new_alert` in loop.py: inherit service / cluster / region
-    # across subsequent investigations in the same REPL session.  Without
-    # this, follow-up free-text alerts would lose context that `/investigate`
-    # just discovered (#243 requirement 7: "the session remembers everything").
-    session.accumulate_from_state(final_state)
-    session.record("alert", f"/investigate {args[0]}")
-    return True
-
-
-def _cmd_list(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    sub = (args[0].lower() if args else "").strip()
-
-    if sub in ("integrations", "integration", "int"):
-        _render_integrations_table(console, _load_verified_integrations())
-        return True
-
-    if sub in ("mcp", "mcps"):
-        _render_mcp_table(console, _load_verified_integrations())
-        return True
-
-    if sub in ("models", "model", "llm", "llms"):
-        _render_models_table(console)
-        return True
-
-    if sub and sub not in ("", "all"):
-        console.print(
-            f"[{TERMINAL_ERROR}]unknown list target:[/] {escape(sub)}  "
-            "(try [bold]/list integrations[/bold], [bold]/list models[/bold], "
-            "or [bold]/list mcp[/bold])"
-        )
-        return True
-
-    # Default: summary view — show everything compactly.
-    results = _load_verified_integrations()
-    _render_integrations_table(console, results)
-    _render_mcp_table(console, results)
-    _render_models_table(console)
-    return True
-
-
-def _cmd_history(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    entries = load_command_history_entries()
-    if not entries:
-        console.print("[dim]no history yet.[/dim]")
-        return True
-
-    table = _repl_table(title="Command history", title_style=TERMINAL_ACCENT_BOLD)
-    table.add_column("#", style="dim", justify="right")
-    table.add_column("text", overflow="fold")
-
-    for i, entry in enumerate(entries, start=1):
-        table.add_row(str(i), escape(entry))
-    console.print(table)
-    return True
-
-
-def _cmd_last(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    if session.last_state is None:
-        console.print("[dim]no investigation in this session yet.[/dim]")
-        return True
-
-    root_cause = session.last_state.get("root_cause", "")
-    report = session.last_state.get("problem_md") or session.last_state.get("slack_message") or ""
-
-    if root_cause:
-        console.print(f"[{TERMINAL_ACCENT_BOLD}]root cause:[/] {escape(str(root_cause))}")
-    if report:
-        console.print(escape(str(report)))
-    if not root_cause and not report:
-        console.print("[dim]last investigation has no report content.[/dim]")
-    return True
-
-
-def _cmd_save(session: ReplSession, console: Console, args: list[str]) -> bool:
-    import json
-    from pathlib import Path
-
-    if session.last_state is None:
-        console.print("[dim]nothing to save — run an investigation first.[/dim]")
-        return True
-
-    if not args:
-        console.print("[dim]usage:[/dim] /save <path>  (e.g. /save report.md or /save out.json)")
-        return True
-
-    dest = Path(args[0])
-    try:
-        if dest.suffix.lower() == ".json":
-            dest.write_text(json.dumps(session.last_state, indent=2, default=str), encoding="utf-8")
-        else:
-            root_cause = session.last_state.get("root_cause", "")
-            report = (
-                session.last_state.get("problem_md")
-                or session.last_state.get("slack_message")
-                or ""
-            )
-            lines = []
-            if root_cause:
-                lines.append(f"## Root Cause\n\n{root_cause}\n")
-            if report:
-                lines.append(f"## Report\n\n{report}\n")
-            dest.write_text("\n".join(lines) or "(no report content)", encoding="utf-8")
-        console.print(f"[green]saved:[/green] {escape(str(dest))}")
-    except Exception as exc:  # noqa: BLE001
-        console.print(f"[{TERMINAL_ERROR}]save failed:[/] {escape(str(exc))}")
-    return True
-
-
-def _cmd_context(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    if not session.accumulated_context:
-        console.print("[dim]no infra context accumulated yet.[/dim]")
-        return True
-
-    table = _repl_table(
-        title="Accumulated context", title_style=TERMINAL_ACCENT_BOLD, show_header=False
-    )
-    table.add_column("key", style="bold")
-    table.add_column("value")
-    for k, v in sorted(session.accumulated_context.items()):
-        table.add_row(k, escape(str(v)))
-    console.print(table)
-    return True
-
-
-def _cmd_cost(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    table = _repl_table(title="Session cost", title_style=TERMINAL_ACCENT_BOLD, show_header=False)
-    table.add_column("key", style="bold")
-    table.add_column("value")
-    table.add_row("interactions", str(len(session.history)))
-
-    if session.token_usage:
-        inp = session.token_usage.get("input", 0)
-        out = session.token_usage.get("output", 0)
-        table.add_row("input tokens", f"{inp:,}")
-        table.add_row("output tokens", f"{out:,}")
-    else:
-        table.add_row("token usage", "[dim]not available (LangSmith not wired yet)[/dim]")
-
-    console.print(table)
-    return True
-
-
-def _cmd_verbose(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    if args and args[0].lower() in ("off", "false", "0", "disable"):
-        os.environ.pop("TRACER_VERBOSE", None)
-        console.print("[dim]verbose logging off[/dim]")
-    else:
-        os.environ["TRACER_VERBOSE"] = "1"
-        console.print("[yellow]verbose logging on[/yellow]")
-    return True
-
-
-def _cmd_compact(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    before = len(session.history)
-    if before > 20:
-        session.history = session.history[-20:]
-        console.print(f"[dim]compacted: kept last 20 of {before} entries.[/dim]")
-    else:
-        console.print(f"[dim]nothing to compact ({before} entries, limit is 20).[/dim]")
-    return True
-
-
-def _task_started_label(task: TaskRecord) -> str:
-    return datetime.fromtimestamp(task.started_at, tz=UTC).strftime("%Y-%m-%d %H:%M:%S UTC")
-
-
-def _task_duration_label(task: TaskRecord) -> str:
-    duration = task.duration_seconds()
-    if duration is None:
-        return "—"
-    return f"{duration:.1f}s"
-
-
-def _task_detail_label(task: TaskRecord) -> str:
-    if task.error:
-        return str(task.error)
-    if task.result:
-        return str(task.result)
-    return "—"
-
-
-def _cmd_tasks(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    # ``list_recent(n)`` respects ``n`; default in the helper is smaller—use 50 here so
-    # /tasks spans most of what the registry can hold (~100 slots).
-    tasks = session.task_registry.list_recent(n=50)
-    if not tasks:
-        console.print("[dim]no tasks recorded this session.[/dim]")
-        return True
-
-    table = _repl_table(title="Tasks", title_style=TERMINAL_ACCENT_BOLD)
-    table.add_column("id", style="bold")
-    table.add_column("kind")
-    table.add_column("status")
-    table.add_column("started", style="dim")
-    table.add_column("duration", style="dim", justify="right")
-    table.add_column("detail", style="dim", overflow="fold")
-
-    status_style = {
-        TaskStatus.RUNNING: "yellow",
-        TaskStatus.COMPLETED: "green",
-        TaskStatus.CANCELLED: "yellow",
-        TaskStatus.FAILED: "red",
-        TaskStatus.PENDING: "dim",
-    }
-    for task in tasks:
-        st = status_style.get(task.status, "dim")
-        table.add_row(
-            task.task_id,
-            task.kind.value,
-            f"[{st}]{task.status.value}[/{st}]",
-            _task_started_label(task),
-            _task_duration_label(task),
-            escape(_task_detail_label(task)),
-        )
-    console.print(table)
-    return True
-
-
-def _cmd_cancel(session: ReplSession, console: Console, args: list[str]) -> bool:
-    if not args:
-        console.print(
-            f"[{TERMINAL_ERROR}]usage:[/] /cancel <task_id>  — use [bold]/tasks[/bold] to list ids"
-        )
-        return True
-
-    needle = args[0]
-    candidates = session.task_registry.candidates(needle)
-    if not candidates:
-        console.print(f"[{TERMINAL_ERROR}]no task matches id:[/] {escape(needle)}")
-        return True
-    if len(candidates) > 1:
-        console.print(
-            f"[{TERMINAL_ERROR}]ambiguous id prefix:[/] {escape(needle)} "
-            f"[dim]({len(candidates)} matches — use a longer prefix)[/dim]"
-        )
-        return True
-
-    task = candidates[0]
-    if task.status != TaskStatus.RUNNING:
-        console.print(
-            f"[dim]task {escape(task.task_id)} already finished "
-            f"(status: {task.status.value}).[/dim]"
-        )
-        return True
-
-    task.request_cancel()
-    if task.kind == TaskKind.INVESTIGATION:
-        console.print(
-            "[yellow]cancellation signaled.[/yellow] "
-            "[dim]if the investigation is still streaming, press[/dim] [bold]Ctrl+C[/bold] "
-            "[dim]to interrupt the current run.[/dim]"
-        )
-    else:
-        console.print(
-            f"[green]stop requested[/green] [dim]for synthetic test {escape(task.task_id)}.[/dim] "
-            "[dim]use[/dim] [bold]/tasks[/bold] [dim]to confirm status.[/dim]"
-        )
-    return True
-
-
-def _cmd_stop(session: ReplSession, console: Console, args: list[str]) -> bool:  # noqa: ARG001
-    console.print(
-        "[dim]in-flight work: press[/dim] [bold]Ctrl+C[/bold] [dim]during a streaming investigation, "
-        "or run[/dim] [bold]/tasks[/bold] [dim]then[/dim] [bold]/cancel <id>[/bold] [dim]for background tasks."
-    )
-    return True
-
-
-SLASH_COMMANDS: dict[str, SlashCommand] = {
-    "/help": SlashCommand("/help", "show available commands", _cmd_help),
-    "/?": SlashCommand("/?", "shortcut for /help", _cmd_help),
-    "/exit": SlashCommand("/exit", "exit the interactive shell", _cmd_exit),
-    "/quit": SlashCommand("/quit", "alias for /exit", _cmd_exit),
-    "/clear": SlashCommand("/clear", "clear the screen and re-render the banner", _cmd_clear),
-    "/reset": SlashCommand("/reset", "clear session state (keeps trust mode)", _cmd_reset),
-    "/trust": SlashCommand("/trust", "toggle trust mode ('/trust off' to disable)", _cmd_trust),
-    "/status": SlashCommand("/status", "show session status", _cmd_status),
-    "/list": SlashCommand(
-        "/list",
-        "list integrations, MCP servers, and the active LLM connection "
-        "('/list integrations', '/list models', '/list mcp')",
-        _cmd_list,
-    ),
-    "/integrations": SlashCommand(
-        "/integrations",
-        "manage integrations ('/integrations list', '/integrations verify', '/integrations show <service>')",
-        _cmd_integrations,
-    ),
-    "/mcp": SlashCommand(
-        "/mcp",
-        "manage MCP servers ('/mcp list', '/mcp connect', '/mcp disconnect')",
-        _cmd_mcp,
-    ),
-    "/model": SlashCommand(
-        "/model",
-        "show or set the active LLM ('/model show', "
-        "'/model set <provider> [model] [--toolcall-model <m>]', "
-        "'/model toolcall set <model>')",
-        _cmd_model,
-    ),
-    "/health": SlashCommand("/health", "show integration and agent health", _cmd_health),
-    "/doctor": SlashCommand("/doctor", "run full environment diagnostic", _cmd_doctor),
-    "/version": SlashCommand("/version", "print version, Python and OS info", _cmd_version),
-    "/template": SlashCommand(
-        "/template",
-        "print a starter alert JSON template ('/template generic|datadog|grafana|honeycomb|coralogix')",
-        _cmd_template,
-    ),
-    "/investigate": SlashCommand(
-        "/investigate",
-        "run an RCA investigation from a file ('/investigate <file>')",
-        _cmd_investigate_file,
-    ),
-    "/history": SlashCommand("/history", "show persisted command history", _cmd_history),
-    "/last": SlashCommand("/last", "reprint the most recent investigation report", _cmd_last),
-    "/save": SlashCommand("/save", "save last investigation to a file ('/save <path>')", _cmd_save),
-    "/context": SlashCommand("/context", "show accumulated infra context", _cmd_context),
-    "/cost": SlashCommand("/cost", "show token usage and session cost", _cmd_cost),
-    "/verbose": SlashCommand(
-        "/verbose", "toggle verbose logging ('/verbose off' to disable)", _cmd_verbose
-    ),
-    "/compact": SlashCommand("/compact", "trim old session history to free memory", _cmd_compact),
-    "/tasks": SlashCommand("/tasks", "list recent and in-flight shell tasks", _cmd_tasks),
-    "/cancel": SlashCommand(
-        "/cancel", "cancel a running task by id ('/cancel <task_id>' — see /tasks)", _cmd_cancel
-    ),
-    "/stop": SlashCommand(
-        "/stop", "hints for stopping in-flight investigations and background tasks", _cmd_stop
-    ),
-}
-
-
-def dispatch_slash(command_line: str, session: ReplSession, console: Console) -> bool:
-    """Dispatch a slash command line. Returns False iff the REPL should exit."""
-    stripped = command_line.strip()
-    if stripped == "/":
-        return _cmd_help(session, console, [])
-
-    parts = stripped.split()
-    if not parts:
-        return True
-    name = parts[0].lower()
-    args = parts[1:]
-    cmd = SLASH_COMMANDS.get(name)
-    if cmd is None:
-        console.print()
-        console.print(
-            f"[{TERMINAL_ERROR}]unknown command:[/] {escape(name)}  (type [bold]/help[/bold])"
-        )
-        return True
-    return cmd.handler(session, console, args)
+from app.cli.interactive_shell.command_registry import (
+    SLASH_COMMANDS,
+    dispatch_slash,
+    switch_llm_provider,
+    switch_toolcall_model,
+)
+from app.cli.interactive_shell.command_registry.types import SlashCommand
+
+__all__ = [
+    "SLASH_COMMANDS",
+    "SlashCommand",
+    "dispatch_slash",
+    "switch_llm_provider",
+    "switch_toolcall_model",
+]
