+++ tests/cli/interactive_shell/test_tasks.py
@@ -195,9 +195,9 @@ class TestSyntheticSubprocessWatcher:
     def test_watch_marks_completed_when_process_already_done(
         self, monkeypatch: pytest.MonkeyPatch
     ) -> None:
-        import app.cli.interactive_shell.agent_actions as aa
+        import app.cli.interactive_shell.action_executor as ae
 
-        monkeypatch.setattr(aa.threading, "Thread", _ImmediateThread)
+        monkeypatch.setattr(ae.threading, "Thread", _ImmediateThread)
 
         proc = MagicMock()
         proc.poll.return_value = 0
@@ -208,7 +208,7 @@ class TestSyntheticSubprocessWatcher:
         task.mark_running()
         task.attach_process(proc)
         stderr_buf: tempfile.SpooledTemporaryFile[bytes] = tempfile.SpooledTemporaryFile()  # type: ignore[type-arg]  # noqa: SIM115
-        aa._watch_synthetic_subprocess(task, proc, session, "rds_postgres", stderr_buf)
+        ae.watch_synthetic_subprocess(task, proc, session, "rds_postgres", stderr_buf)
         assert task.status == TaskStatus.COMPLETED
         hist = session.history[-1]
         assert hist["type"] == "synthetic_test"
@@ -224,9 +224,9 @@ class TestSyntheticSubprocessWatcher:
         cancel_requested branch runs, so terminated_by_watcher stays False.
         The task must be COMPLETED, not CANCELLED — the process succeeded.
         """
-        import app.cli.interactive_shell.agent_actions as aa
+        import app.cli.interactive_shell.action_executor as ae
 
-        monkeypatch.setattr(aa.threading, "Thread", _ImmediateThread)
+        monkeypatch.setattr(ae.threading, "Thread", _ImmediateThread)
 
         session = ReplSession()
         task = session.task_registry.create(TaskKind.SYNTHETIC_TEST)
@@ -250,9 +250,9 @@ class TestSyntheticSubprocessWatcher:
             task.cancel_requested.set()
             pending[0] = 0  # process finishes naturally in the same window
 
-        monkeypatch.setattr(aa.time, "sleep", _fake_sleep)
+        monkeypatch.setattr(ae.time, "sleep", _fake_sleep)
         stderr_buf: tempfile.SpooledTemporaryFile[bytes] = tempfile.SpooledTemporaryFile()  # type: ignore[type-arg]  # noqa: SIM115
-        aa._watch_synthetic_subprocess(task, proc, session, "rds_postgres", stderr_buf)
+        ae.watch_synthetic_subprocess(task, proc, session, "rds_postgres", stderr_buf)
         # terminated_by_watcher is False → honour exit code 0 → COMPLETED
         assert task.status == TaskStatus.COMPLETED
         assert sleeps
@@ -261,9 +261,9 @@ class TestSyntheticSubprocessWatcher:
         self, monkeypatch: pytest.MonkeyPatch
     ) -> None:
         """cancel_requested is set while proc is still running; watcher terminates it."""
-        import app.cli.interactive_shell.agent_actions as aa
+        import app.cli.interactive_shell.action_executor as ae
 
-        monkeypatch.setattr(aa.threading, "Thread", _ImmediateThread)
+        monkeypatch.setattr(ae.threading, "Thread", _ImmediateThread)
 
         session = ReplSession()
         task = session.task_registry.create(TaskKind.SYNTHETIC_TEST)
@@ -278,10 +278,10 @@ class TestSyntheticSubprocessWatcher:
         task.cancel_requested.set()  # cancel already set before first loop check
 
         # Skip the sleep so the loop iterates immediately to the cancel branch.
-        monkeypatch.setattr(aa.time, "sleep", lambda _: None)
+        monkeypatch.setattr(ae.time, "sleep", lambda _: None)
 
         stderr_buf: tempfile.SpooledTemporaryFile[bytes] = tempfile.SpooledTemporaryFile()  # type: ignore[type-arg]  # noqa: SIM115
-        aa._watch_synthetic_subprocess(task, proc, session, "rds_postgres", stderr_buf)
+        ae.watch_synthetic_subprocess(task, proc, session, "rds_postgres", stderr_buf)
         assert task.status == TaskStatus.CANCELLED
         hist = session.history[-1]
         assert hist["type"] == "synthetic_test"
@@ -295,9 +295,9 @@ class TestSyntheticSubprocessWatcher:
         The watcher should mark the task COMPLETED, not CANCELLED, because we
         never called _terminate_child_process — the process was already gone.
         """
-        import app.cli.interactive_shell.agent_actions as aa
+        import app.cli.interactive_shell.action_executor as ae
 
-        monkeypatch.setattr(aa.threading, "Thread", _ImmediateThread)
+        monkeypatch.setattr(ae.threading, "Thread", _ImmediateThread)
 
         session = ReplSession()
         task = session.task_registry.create(TaskKind.SYNTHETIC_TEST)
@@ -312,14 +312,14 @@ class TestSyntheticSubprocessWatcher:
         task.cancel_requested.set()
 
         stderr_buf: tempfile.SpooledTemporaryFile[bytes] = tempfile.SpooledTemporaryFile()  # type: ignore[type-arg]  # noqa: SIM115
-        aa._watch_synthetic_subprocess(task, proc, session, "rds_postgres", stderr_buf)
+        ae.watch_synthetic_subprocess(task, proc, session, "rds_postgres", stderr_buf)
         assert task.status == TaskStatus.COMPLETED
 
     def test_watch_captures_stderr_on_failure(self, monkeypatch: pytest.MonkeyPatch) -> None:
         """Diagnostic stderr output is included in mark_failed message."""
-        import app.cli.interactive_shell.agent_actions as aa
+        import app.cli.interactive_shell.action_executor as ae
 
-        monkeypatch.setattr(aa.threading, "Thread", _ImmediateThread)
+        monkeypatch.setattr(ae.threading, "Thread", _ImmediateThread)
 
         session = ReplSession()
         task = session.task_registry.create(TaskKind.SYNTHETIC_TEST)
@@ -330,7 +330,7 @@ class TestSyntheticSubprocessWatcher:
 
         stderr_buf: tempfile.SpooledTemporaryFile[bytes] = tempfile.SpooledTemporaryFile()  # type: ignore[type-arg]  # noqa: SIM115
         stderr_buf.write(b"ConnectionError: database unreachable\n")
-        aa._watch_synthetic_subprocess(task, proc, session, "rds_postgres", stderr_buf)
+        ae.watch_synthetic_subprocess(task, proc, session, "rds_postgres", stderr_buf)
         assert task.status == TaskStatus.FAILED
         assert "exit code 1" in (task.error or "")
         assert "ConnectionError" in (task.error or "")
@@ -338,10 +338,10 @@ class TestSyntheticSubprocessWatcher:
     def test_watch_skips_synthetic_history_after_reset(
         self, monkeypatch: pytest.MonkeyPatch
     ) -> None:
-        import app.cli.interactive_shell.agent_actions as aa
+        import app.cli.interactive_shell.action_executor as ae
 
         _DeferredSyntheticThread.pending.clear()
-        monkeypatch.setattr(aa.threading, "Thread", _DeferredSyntheticThread)
+        monkeypatch.setattr(ae.threading, "Thread", _DeferredSyntheticThread)
 
         proc = MagicMock()
         proc.poll.return_value = 0
@@ -352,7 +352,7 @@ class TestSyntheticSubprocessWatcher:
         task.mark_running()
         task.attach_process(proc)
         stderr_buf: tempfile.SpooledTemporaryFile[bytes] = tempfile.SpooledTemporaryFile()  # type: ignore[type-arg]  # noqa: SIM115
-        aa._watch_synthetic_subprocess(task, proc, session, "rds_postgres", stderr_buf)
+        ae.watch_synthetic_subprocess(task, proc, session, "rds_postgres", stderr_buf)
         assert len(_DeferredSyntheticThread.pending) == 1
         session.clear()
         _DeferredSyntheticThread.pending[0]()
@@ -362,10 +362,10 @@ class TestSyntheticSubprocessWatcher:
     def test_deferred_watcher_writes_history_when_no_reset(
         self, monkeypatch: pytest.MonkeyPatch
     ) -> None:
-        import app.cli.interactive_shell.agent_actions as aa
+        import app.cli.interactive_shell.action_executor as ae
 
         _DeferredSyntheticThread.pending.clear()
-        monkeypatch.setattr(aa.threading, "Thread", _DeferredSyntheticThread)
+        monkeypatch.setattr(ae.threading, "Thread", _DeferredSyntheticThread)
 
         proc = MagicMock()
         proc.poll.return_value = 0
@@ -376,7 +376,7 @@ class TestSyntheticSubprocessWatcher:
         task.mark_running()
         task.attach_process(proc)
         stderr_buf: tempfile.SpooledTemporaryFile[bytes] = tempfile.SpooledTemporaryFile()  # type: ignore[type-arg]  # noqa: SIM115
-        aa._watch_synthetic_subprocess(task, proc, session, "rds_postgres", stderr_buf)
+        ae.watch_synthetic_subprocess(task, proc, session, "rds_postgres", stderr_buf)
         _DeferredSyntheticThread.pending[0]()
         assert session.history[-1]["type"] == "synthetic_test"
         _DeferredSyntheticThread.pending.clear()
